+++ ChangeLog.md
@@ -5,8 +5,21 @@ All notable changes to this project will be documented in this file.
 
 
 
-# Release 3.23 - 18/12/2023
+# Release 3.24 - 08/04/2024
+- FIX : Module_number missed in subTotal class. function  addSubTotalLine function  in test order_supplier - *26/06/2024* - 3.24.4  
+- FIX : display totalht ligne 'non compris' - *12/04/2024* - 3.24.3  
+- FIX : Title extrafields wasn't working - *12/04/2024* - 3.24.2  
+- FIX : doublon affichage label description lors de la création de facture depuis un objet (propal/commande ...)  - *03/04/2024* - 3.24.1  
+- NEW : Ajout des quantités par sous totaux sur l'interface - *27/03/2024* - 3.24.0
 
+# Release 3.23 - 18/12/2023
+- FIX : DA024939 - Added static method hasBreakPage to check if a line has a break page or not - *17/05/2024* - 3.23.10
+- FIX : DA024587 - Les totaux remisés sur le PDF Sponge sont erronés  - *20/03/2024* - 3.23.9
+- FIX : Suite a l'issue #379, la création de facture d'acompte avec un montant variable change les qté des lignes générées par le module donc on utilise un trigger pour remettre les bonnes qtés - *18/03/2024* - 3.23.8
+- FIX : Sur le modèle PDF Sponge, retrait des lignes de sous totaux au moment du calcul de l'avancement global - *12/03/2024* - 3.23.7
+- FIX : Create invoice deposit line same order with percent : FIX add Line subtotal and free text  - *20/02/2024* - 3.23.6
+- FIX : Display subtotal with tax when global PDF_PROPAL_SHOW_PRICE_INCL_TAX is enable - *15/01/2024* - 3.23.5
+- FIX : php 8.2 warning - *15/01/2024* - 3.23.4
 - FIX : colum in card if global 'MAIN_NO_INPUT_PRICE_WITH_TAX' is enable
 - FIX : js error when summary menu is disabled - *06/10/2023* - 3.23.2
 - FIX : css break due to incompatible conf - *06/10/2023* - 3.23.1
@@ -19,6 +32,7 @@ Changed Dolibarr compatibility range to 15 min - 19 max
 
 # Release 3.22 - 19/07/2023
 
+- FIX : DA024364 - Fatal error deuxieme parametre de str_repeat - *19/01/2024* - 3.22.7
 - FIX : DA024159 - Suppression des ligne sous total hors bloc expédié dans le bon d'expédition - *13/12/2023* - 3.22.6
 - HOTFIX : must be greater than 0 fatal - *22/11/2023* - 3.22.5
 - FIX : DA024057 - Anomalie PDF lors de l'activation conf "Remplacer par le détail des TVA si l'option "Cacher le détail des ensembles" est utilisée (expérimental)" - *16/11/2023* - 3.22.4